Enable `-Wdouble-promotion` compiler warning
Fix implicit float-to-double and double-to-long-double promotions:
- Add `L` suffix to `long double` array literals in `shift10` and
  `BloomFilterHash` that were missing it
- Add explicit `static_cast<double>` for `Float32` settings used in
  `double` contexts (comparisons, `std::bernoulli_distribution`, etc.)
- Replace `NAN`/`INFINITY` (which are `float`) with
  `std::numeric_limits<double>` equivalents in `preciseExp10`
- Suppress the warning with pragmas in template-heavy generic math
  code where float operations are intentionally done at float
  precision (aggregate functions, array functions, etc.)

Reference: https://github.com/ClickHouse/ClickHouse/issues/76944

+++ programs/server/Server.cpp
@@ -2289,7 +2289,7 @@ try
 
             global_context->setS3QueueDisableStreaming(new_server_settings[ServerSetting::s3queue_disable_streaming]);
 
-            global_context->setOSCPUOverloadSettings(new_server_settings[ServerSetting::min_os_cpu_wait_time_ratio_to_drop_connection], new_server_settings[ServerSetting::max_os_cpu_wait_time_ratio_to_drop_connection]);
+            global_context->setOSCPUOverloadSettings(static_cast<double>(new_server_settings[ServerSetting::min_os_cpu_wait_time_ratio_to_drop_connection]), static_cast<double>(new_server_settings[ServerSetting::max_os_cpu_wait_time_ratio_to_drop_connection]));
 
             size_t remote_read_bandwidth = new_server_settings[ServerSetting::max_remote_read_network_bandwidth_for_server];
             size_t remote_write_bandwidth = new_server_settings[ServerSetting::max_remote_write_network_bandwidth_for_server];
@@ -2338,7 +2338,7 @@ try
             {
                 const auto & new_pool_size = new_server_settings[ServerSetting::background_pool_size];
                 const auto & new_ratio = new_server_settings[ServerSetting::background_merges_mutations_concurrency_ratio];
-                global_context->getMergeMutateExecutor()->increaseThreadsAndMaxTasksCount(new_pool_size, static_cast<size_t>(static_cast<double>(new_pool_size) * new_ratio));
+                global_context->getMergeMutateExecutor()->increaseThreadsAndMaxTasksCount(new_pool_size, static_cast<size_t>(static_cast<double>(new_pool_size) * static_cast<double>(new_ratio)));
                 global_context->getMergeMutateExecutor()->updateSchedulingPolicy(new_server_settings[ServerSetting::background_merges_mutations_scheduling_policy].toString());
             }
